remove warning on WIN32
git-svn-id: svn://svn.savannah.nongnu.org/osip/trunk/osip@172 a617929b-4c5a-41d8-b177-2eda7db2ca6c

+++ parser/msg_read.c
@@ -317,7 +317,7 @@ strcat_headers_all_on_one_line (char **_string, int *malloc_size,
 	      *next = NULL;
 	      return -1;
 	    }
-	  if (*malloc_size < message - string + strlen (tmp) + 100)
+	  if (*malloc_size < message - string + (int)strlen (tmp) + 100)
 	    {
 	      int size = message - string;
 	      *malloc_size = message - string + (int)strlen (tmp) + 100;
